Added rtruediv and pow operators to Quantity class
